# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! 
    You can upload your photo directly from your PC, and they will be visible to everyone following you
  version: 1.0.0

#============================================TAGS PART==============================================#
...
security:
  - bearerAuth: []
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Bearer token for the WASAPhoto API.
      scheme: bearer

#============================================SCHEMAS PART==============================================#
  schemas:
...
    Code-Identifier:
      title: Code-random
      type: object
      description: |-
        General random generated string, which will be used as a unique identifier.
        It is a unique code that identifier uniquely:
          - the user
          - the photo
          - the comment
      properties:
        identifier: 
          type: string
          description: unique identifier generatedr randomly
          pattern: '^.*?$'
          minLength: 64
          maxLength: 64
          example: "115981bedf5cc6956330a1e4c394526a8839e8a3ec3adf0fre3eedd59c759a58"
      
    
#_____________________________USER FIELD____________________________________________
#-------------------------------------------------------------------------------- 
...
    UserID:
      title: UserID
      description: It's an object that contains an identifier code of user
      type: object
      properties:
        code-user: 
          $ref: "#/components/schemas/Code-Identifier"
      
#--------------------------------------------------------------------------------
...
    Error:
      title: Error
      type: object
      description: |-
        An error message that contains:
          - Code 
          - Message
      properties:
        code-error:
          type: integer
          description: The error code
          minimum: 3
          maximum: 3

        message-error:
          type: string
          description: The error message
          minLength: 1
          maxLength: 256
          pattern: ^[^\\]{1,256}$

#============================================PARAMETERS PART==============================================#
  parameters:
    user-id:
      name: user-id
      in: path
      description: The ID of a user to get information about him/her
      required: true
      schema: 
        $ref: "#/components/schemas/Username/properties/username-string"
      
#--------------------------------------------------------------------------------
...
    following-id:
      name: following-id
      in: path
      description: The ID of a following to get information about him/her
      required: true
      schema: 
        $ref: "#/components/schemas/Username/properties/username-string"

#--------------------------------------------------------------------------------
...
paths:
...
  /users/{user-id}/profile/following/{following-id}:
    parameters: 
      - $ref: "#/components/parameters/user-id"
      - $ref: "#/components/parameters/following-id"
  
    put:
      tags: ["follower"]
      summary: Follow User
      description: the user follow a photo so the user must be authenticated
      operationId: followUser
      
      requestBody:
        description: the follow user id
        required: true
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/UserID"
            
      responses:
        "204":
          description: "The user has been follow"

        "400":
          description: Bad Request
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "401":
          description: |-
            The following-id or requested-id is invalid, no authentication information
            has been provided
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "403":
          description: The following-id is forbidden
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "404":
          description: |-
           Resource is not Found: either the requested user or following user does not exist
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "500":
          description: Internal Server Error
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"


    delete:
      tags: ["follower"]
      summary: Unban User
      description: the user unfollow a photo so the user must be authenticated
      operationId: unfollowUser
      
      responses:
        "204":
          description: "The user has been unfollow"

        "400":
          description: Bad Request
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "401":
          description: |-
            The following-id is invalid, no authentication information
            has been provided
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "403":
          description: The following-id is forbidden
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "404":
          description: |-
            Resource is not Found: either the requested user or following user does not exist
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"

        "500":
          description: Internal Server Error
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Error"


  
#_____________________________PROFILE - GALLERY _________________________
#----------------------------------------------------------------------
...